requirements.txt
Flask
flask_sqlalchemy
flask_login
flask_wtf
email_validator
bcrypt
cryptography
python-dotenv